drastically speed up line-numbering by better ordering of operations
all browsers appear to cache layout information (offsetHeight and such), and recompute
it when the dom is changed in any way. the line-numbering used to add number nodes
one-at-a-time, checking whether the offsetHeight of the number div was high enough
after every number. this is unneccesarily slow -- computing how many numbers we need,
and then adding them all at once (and a similar scheme for making the gaps when
text-wrapping is on) sped things up a *lot*.