Go back to webkitLastLineHack
The proposed solution (using execCommand("insertLineBreak") on webkit)
was no real solution -- all that does is insert to linebreaks when at
the end of the document. (This is the webkit code working around its
own bugs.) There are other situations (such as pasting or importing
code) that can cause the last-line bug to manifest itself.

This brings back most of the old webkitLastLineHack code, but uses an
extra BR (with a hackBR property) instead of a zero-width space. This
seems to cause less issues. getCode has been altered to handle this
spurious extra BR, but other code has not (so the number of lines
reported, for example, will still be one off).