prevent spurious re-focusing in IE
some of the editor methods only work on IE when the editor is focused, so
they were focusing it if the browser is IE. However, due to a recent
optimization, this sometimes caused cursor-jumping when the user was typing
in the editor. It now first checks whether the editor isn't already
focused.